[PORT FROM MAIN]WiFi: Disable Wifi sleep policy when USB tethering is on
BZ: 15926 17650

The wifi sleep policy "When screen turns off" will switch off wifi although the USB tethering is active.
With this patch, the sleep policy is removed when the usb tethering is active.
The WifiService maintains a local status variable in according with the wifi tethering (on,off).
If the usb tethering is on, the alarm, to switch off the wifi, is cancelled.

Change-Id: I516cac357078fd711c1d0a63bc5c9fe526027060
Orig-Change-Id: I828e5f5c93d94be9ac03f06fe6ed07c1eb17aa81
Signed-off-by: Marco Sinigaglia <marcox.sinigaglia@intel.com>
Signed-off-by: jonathan.de.cesco <jonathan.de.cesco@intel.com>
Reviewed-on: http://android.intel.com:8080/80662
Reviewed-by: Lemetayer, Jean-marieX <jean-mariex.lemetayer@intel.com>
Reviewed-by: Garcia, JeremieX <jeremiex.garcia@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>