[PORT FROM MAIN] WLAN: remove buffer limitation from wifi JNI command
BZ: 43667

Some doxxxCommand use an internal buffer of 256 bytes long.
In case of big command string to build, this buffer is not big enough
and the doxxxCommand return an error.

This could be avoided by using a dynamic allocation.
This fix keep the original buffer (256 bytes long) for a first trial.
In the case the command string to build is too big, allocate a bigger enough
buffer and use it. It's a good compromise for performance.

Change-Id: I919932557fd8f7767138bdc3e20c17dccca5927f
Signed-off-by: Honore Tricot <honorex.tricot@intel.com>
Reviewed-on: http://android.intel.com:8080/80303
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>