To fix the green title edge problem (JNI)
BZ: 39058

The problem was caused by alpha blending between some useless
green lines in the title image and the input image. In the
patch, we correct this wrong title boundary and only perform
blending between input image and title area.
There is also a very great side effect that: the computation
complexity is significantly reduced. It saves the exporting
time up to 60% for 1080P video with title effect.

Change-Id: I65738f495b78231dfa5327e056c2584035cdc260
Signed-off-by: wfeng6 <wei.feng@intel.com>
Reviewed-on: http://android.intel.com:8080/75047
Reviewed-by: lab_cactus <lab_cactus@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Ding, Haitao <haitao.ding@intel.com>
Tested-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: Tong, BoX <box.tong@intel.com>