[PORT FROM MAIN] To fix the green title edge problem (JNI)
BZ: 39058

The problem was caused by alpha blending between some useless
green lines in the title image and the input image. In the
patch, we correct this wrong title boundary and only perform
blending between input image and title area.
There is also a very great side effect that: the computation
complexity is significantly reduced. It saves the exporting
time up to 60% for 1080P video with title effect.

Change-Id: I93b310c7f9bf73109b1889b5c83c30aaf4ec8aba
Orig-Change-Id: I65738f495b78231dfa5327e056c2584035cdc260
Signed-off-by: wfeng6 <wei.feng@intel.com>
Signed-off-by: Gu, Wangyi <wangyi.gu@intel.com>
Reviewed-on: http://android.intel.com:8080/121765
Reviewed-by: Liang, Dan <dan.liang@intel.com>
Reviewed-by: Ding, Haitao <haitao.ding@intel.com>
Tested-by: Ding, Haitao <haitao.ding@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>