[PORT FROM MAIN]Audio response abnormally when disable HDMI connection.
BZ: 50214

Original BZ: 38559, 42604

Patch 1
When HDMI is enabled/disabled from settings application,
it sends an intent HDMI_SET_STATUS and accordingly
updates state of HDMIEnable. In AudioService when this
intent is received, depending on state of HDMIEnable
audio routing decision is taken. Hence 2 states will be
checked before taking this decision. One state is to
check plug in/out state of HDMI cable and other state
is to check enable/disable in settings application.

Change-Id: Ic4720847b77aa3e09f9c057f906a59742040714f
Orig-Change-Id: I3c1f74c94924c53532e368697ccfb6734aaaec70
Signed-off-by: Arulselvan M <arulselvan.m@intel.com>
Reviewed-on: http://android.intel.com:8080/125056
Reviewed-by: Jayanti, Satya Charitardha <satya.charitardha.jayanti@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>