Avoid memory corruption due to Parcel Recycling
BZ: 47466

While processing batterystatsfile, in present code there is a race where a parcel
might be recycled while it is being written to disk in another thread.
(Original Bug 19426: Reboot-loop due to /data/system/batterystats.bin corruption)

The write process is done in 2 steps:
- write into a Parcel (memory block)
- then write this Parcel to disk using Journaled API.
The Journaled API ensures robustness against file-system issues due to power cut.
The writing is done in batch using a thread running commitPendingDataToDisk().
This function writes the parcel to disk but this parcel is a global variable
initialized by writeLocked(). If the writing takes too long time and two write
threads are launched, the parcel is freed by the second writeLocked() call;
which could cause the memory corruption.

Hence, change global Parcel variable by a Parcel parameter to commitPendingDataToDisk()
and only this function will be allowed to recycle the parcel.

Change-Id: I0e76441399f1ce74b83566edb8cf4f977dce58bf
Signed-off-by: Pavan Kumar S <pavan.kumar.s@intel.com>
Reviewed-on: http://android.intel.com:8080/75751
Reviewed-by: lab_cactus <lab_cactus@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: G S, RajkumarX <rajkumarx.g.s@intel.com>
Reviewed-by: Govindaiah, VinayX <vinayx.govindaiah@intel.com>
Tested-by: Kallappa Manjanna, MadhukumarX <madhukumarx.kallappa.manjanna@intel.com>
Reviewed-by: Mariappan, Bhagat SinghX <bhagat.singhx.mariappan@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>