WiFi: WEP auth error due to too small passwords
BZ: 27189

Scenario:
1- We are connected to an AP.
2- Try to connect to a WEP AP with a password of 3 characters
Expected results:
We fail to authenticate to the WEP AP and get back to the AP
we were previously connected.
Obtained results:
We indefinitely try to connect to WEP AP.
The AP we were previously connected on is disabled.

Explanations:
Driver refuses WEP connection with too low password and returns in error
to wpa_supplicant. In reaction, this one raises a message
"SME: Authentication request to the driver failed".
However, this message is not processed by Wifi frameworks and thus, no
counter action is performed.
This patch makes Wifi framework processes this messages using the same
code as for WPA wrong passphrase case.

Change-Id: I1d1842ff4fcc71884f07ca3dff41a12fd2cab076
Orig-Change-Id: I929aa639095ca31a9d28c00e5ee0aea8b90b098d
Signed-off-by: Jeremie GARCIA <jeremiex.garcia@intel.com>
Signed-off-by: jonathan.de.cesco <jonathan.de.cesco@intel.com>
Reviewed-on: http://android.intel.com:8080/83037
Reviewed-by: Bondier, CedricX <cedricx.bondier@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>