WiFi: [BG scan][TI ONLY] Background Scan algorithm activation in Framework.
BZ: 27467 31787 58570

DO NOT UPSTREAM

Enable the INTEL Background Scan by default in the Android Framework.
Do not confuse with the native ICS Backgound Scan which is disabled by default
in the framework(keep it still disbaled).

Following default config for Background learn algorithm is set in the Framework :
short interval : 10 s.
Threshold : -74 dBm
long interval : 60 s.

When, User selects a SSID in the UI WiFi Settings, then
SET_NETWORK <NET_ID> <DEFAULT_BG_SCAN_CONFIG> is sent by Framework to wpa_supplicant
It allows to launch bg scan learn module on the connected SSID in wpa_supplicant.
Then, Background scan algorithm is managed in wpa_supplicant. BG scan
module is stopped by wpa_supplicant itself when user disconnects to network or user
chooses to forget the network.
Hence, the INTEL BG scan does make sense and is still active only in CONNECTED mode.
In addition, BG scan is always running in S0 power mode, WiFi CONNECTED mode if wpa_supplicant process
is running (no BG scan in S3 ).
It means BG scan running when DISPLAY ON and DISPLAY OFF, in and out of the
WiFi Settings Menu.

Today, there is no wakelock in Framework when a BG scan has been
issued by wpa_supplicant. This can lead to a BG scan abort when
platform goes in S3 while wpa_s BG scanning.

Patch proposed prevents going into S3 while BG scan:
1.In Connected state, in case a BGSCAN_LEARN_EVENT
is raised by wpa_s (occuring when issuing a BG scan):
-Acquire a wakelock with a timeout of 2.5 s. (maximal
duration of a scan).
-send command ap_scan=1 in order to authorize a potential
roaming on a better candidate.
2.Upon SCAN_RESULTS_EVENT and in case of BG scan only:
-release the Wakelock above
-Take a Wakelock of 0.5s (maximal roaming duration)
to ensure that Platform does not go in
S3 if a roaming is started by wpa_s.

Test Case ran to validate:
1. Low RSSI triggered by FW while in S3:
a.wakes up the Platform and roams on a better candidate if any.
b.wakes up the Platform and does not roam on a lower candidate

2.BG scan when going out of S3
==> Scan is no more aborted now

3.Roaming when going out of S3 (with candidate higher than TH)
==> roaming OK without going in S3.

hide added bgscan attribute from API
bgscan attribute was added as a public attribute
into WifiConfiguration class but not as hidden.
This is not compliant with android CDD.
Just have to hide this attribute in order to be clean.

Change-Id: I80c43db689f0f3b081f4d23fd59e7d8c0ca4e399
Orig-Change-Id: Ic90732ccae7b87764fd32a697e814365e7e0ab24
Orig-Change-Id: I9d872a5ac5018570601deebcdb3d636c85b32867
Orig-Change-Id: If657d9f9325268bbd2d3432cb6c0bf8e5fd7fec6
Signed-off-by: Jean-Michel Bachot <jean-michel.bachot@intel.com>
Reviewed-on: http://android.intel.com:8080/83054
Reviewed-by: Naoumenko, PaulX <paulx.naoumenko@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>