[PORT FROM MAIN] [VIBRA] Add loading of the vibrator hardware module.
BZ: 49760

Vibrator HAL is now formed of two hardware modules : one is the
default AOSP and the other is the Intel vendor implementation.
The vendor implementation will be loaded prior to the default one.
For that, the vibrator service has to load the module before
registering itself to the JNI.

Change-Id: Id26f54bd4431ef67de544d20eeebba006c74db68
Orig-Change-Id: Id7b1844f043a05dfb58e7d2eda313c0419664fed
JB-integ-Change-Id: I4f01250dee58bb8dffc61c3920957fd4b46fcf1b
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>
Reviewed-on: http://android.intel.com:8080/75593
Reviewed-by: lab_cactus <lab_cactus@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>