[PORT FROM MAIN] [VIBRA] Add loading of the vibrator hardware module.
BZ: 49760 126632

Vibrator HAL is now formed of two hardware modules : one is the
default AOSP and the other is the Intel vendor implementation.
The vendor implementation will be loaded prior to the default one.
For that, the vibrator service has to load the module before
registering itself to the JNI.

Orig-Change-Id: Id26f54bd4431ef67de544d20eeebba006c74db68

Change-Id: I135b317d91af916eb0a8f5d5f4507f24d67ae7b3
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>
Reviewed-on: http://android.intel.com:8080/123540
Reviewed-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>