Vibra: Add loading of the vibrator hardware module.
Vibrator HAL is now formed of two hardware modules: one is the default
AOSP and the other is the vendor implementation.

The vendor implementation will be loaded prior to the default one.
For that, the vibrator service has to load the module before
registering itself to the JNI.

This change is related to other changes in:
 - hardware/libhardware
 - hardware/libhardware_legacy

Change-Id: I844279f5535289f079d412fdc44c5cb3c9c1130c
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-10085
Category: aosp improvement
Domain: AudioComms-HwModules-Vibrator
Origin: upstream https://android-review.googlesource.com/49230
Upstream-Candidate: yes
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>
Signed-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Signed-off-by: David Wagner <david.wagner@intel.com>
Reviewed-on: https://android.intel.com:443/265204
Reviewed-on: https://android.intel.com:443/304002