[PORT FROM MAIN] Add loading of the vibrator hardware module.
BZ: 49760 126632

Vibrator HAL is now formed of two hardware modules : one is the
default AOSP and the other is the Intel vendor implementation.
The vendor implementation will be loaded prior to the default one.
For that, the vibrator service has to load the module before
registering itself to the JNI.

Orig-Change-Id: I135b317d91af916eb0a8f5d5f4507f24d67ae7b3
Category: aosp improvement
Domain: AudioComms-HwModules-Vibrator
Origin: internal
Upstream-Candidate: yes
Change-Id: I85661ee4c934e739425eb117cbbcd8e63ba81d42
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>