[PORT FROM MAIN] Enable Video Editor NV12 Processing on JB (JNI)
BZ: 51587

In video editor's original design, only IYUV420 processing is
supported. But the video format of Intel HW encoder and decoder
is NV12, we have to convert NV12 to IYUV420 on decoder side,
and IYUV420 to NV12 on encoder side. This is one bottleneck
for video editor, which slows down the video export process.

In this patch, we add NV12 support in video editor codes, which
eliminate the color conversion process. This significantly
accelerate the video export.

DO NOT UPSTREAM!

Change-Id: I07197ee5298bb0441d577d3379d97c0a7f9b6867
Orig-Change-Id: Ieae5473690a5391ae4567acec629bc4c410d3b7a
Signed-off-by: wfeng6 <wei.feng@intel.com>
Signed-off-by: Gu, Wangyi <wangyi.gu@intel.com>
Reviewed-on: http://android.intel.com:8080/121766
Reviewed-by: Liang, Dan <dan.liang@intel.com>
Reviewed-by: Ding, Haitao <haitao.ding@intel.com>
Tested-by: Ding, Haitao <haitao.ding@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>