[PORT FROM MAIN] Enable Video Editor NV12 Processing on JB (JNI)
BZ: 51587

In video editor's original design, only IYUV420 processing is
supported. But the video format of Intel HW encoder and decoder
is NV12, we have to convert NV12 to IYUV420 on decoder side,
and IYUV420 to NV12 on encoder side. This is one bottleneck
for video editor, which slows down the video export process.

In this patch, we add NV12 support in video editor codes, which
eliminate the color conversion process. This significantly
accelerate the video export.

DO NOT UPSTREAM!

Change-Id: I4292f129e376e5c5080d4914ac8493a2fba259ff
Orig-Change-Id: I07197ee5298bb0441d577d3379d97c0a7f9b6867
Category: aosp improvement
Domain: Video.Media-video editor
Origin: internal
Upstream-Candidate: no
Signed-off-by: wfeng6 <wei.feng@intel.com>
Signed-off-by: Gu, Wangyi <wangyi.gu@intel.com>
Signed-off-by: chuzhoux <chux.zhou@intel.com>