EM/BCU: Port CurrentMgmtService to JB
BZ: 53564
R4-Mainline
http://android.intel.com:8080/#/c/63112/
[DO NOT UPSTREAM]

This patch ports the CurrentMgmtService to JB.
The SystemServer has to start this service as well, hence
the code change in SystemServer.java.

On receiving an ACTION_BATTERY_CHANGED intent, for a change in the
battery level, this service configures the BCU(Burst Control Unit)
registers, through the sysfs exposed by the BCU driver, to take
some action which would reduce the current consumption.

Change-Id: Iac9da9437c9b63d879cb542294eac2cc491f42bd
Signed-off-by: Pavan Kumar S <pavan.kumar.s@intel.com>
Reviewed-on: http://android.intel.com:8080/75390
Reviewed-by: lab_cactus <lab_cactus@intel.com>
Reviewed-by: Pallala, Ramakrishna <ramakrishna.pallala@intel.com>
Tested-by: Kallappa Manjanna, MadhukumarX <madhukumarx.kallappa.manjanna@intel.com>
Reviewed-by: G S, RajkumarX <rajkumarx.g.s@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>