services_fw: Fix issue in getting cell location
BZ: 50480

Currently, instead of using the cell location framework
APIs, phone.getCellLocation is used to get the
cell location information. This is not working for
all the cases.

With this patch, CellLocation.requestLocationUpdate will
be called to enable location updates. This will
result in onCellLocationChanged notification whenever
there is a change in cell location. If the cell location
is requested, then the received cell location information
will be updated calling native_agps_set_ref_location_cellid.

Change-Id: I8977d848251757652b6a82143b1e0b349677b751
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Fabien Peix <fabien.peix@intel.com>
Reviewed-on: http://android.intel.com:8080/76974
Reviewed-by: Roques, CyrilX <cyrilx.roques@intel.com>
Reviewed-by: Pantaloni, JeromeX <jeromex.pantaloni@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Suet, Nicolas <nicolas.suet@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>