[PORT FROM MAIN] WLAN: Do not convert enterprise password field to string
BZ: 59095

The toString method of WifiConfiguration class permits
to output the EAP password. This could be considered as
a security hole when it is used in WifiConfigController
to generate logs.

The password field is declared as @hide, so there is no reason to
output it.

Fixed by replacing password value by '*' when converted to string.

Change-Id: I5910ef73f1fbe80087d5ad2716b596bbb1efe514
Signed-off-by: Honore Tricot <honorex.tricot@intel.com>
Reviewed-on: http://android.intel.com:8080/80275
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Garcia, JeremieX <jeremiex.garcia@intel.com>
Reviewed-by: De Cesco, Jonathan <jonathan.de.cesco@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>