wifi: Unable to reconnect to a Dual Band AP after turn off/on
BZ: 61500

DO NOT UPSTREAM

It was impossible to reconnect to a Dual Band AP after switch off/on
After turning off AP:
Framework disables network at second authentication failure on the same SSID. This is expected.
First authentication failure happens when phone attempts to roam on 5 GHz band.This is expected.
Then, wpa_s switches into DISCONNECTED and launches a scan: BSSID on 2.4 GHz is sill present in scan results raised by MAC80211 although AP is off.
Indeed, MAC80211 keeps the BSSID during 15 seconds after the beacon loss received.
Second authentication failure happens when phone attempts to reconnect on this 2.4G BSSID.
As a consequence, framework disables the Network SSID
After turning on AP:
As SSID is disabled, wpa_s skips it and automatic reconnection to this SSID is not performed

Workaround here is to increase the number of authentication failures authorized from 2 to 3 in framework. It will prevent the Network SSID disabling in
case of Dual Band AP.

Change-Id: Ifef342280240233d2da158200ccb8dee72dd681c
Orig-Change-Id: I29c8b7c24227ea616fc6f4be0bf373db03cd886b
Signed-off-by: Jean-Michel Bachot <jean-michel.bachot@intel.com>
Reviewed-on: http://android.intel.com:8080/83281
Reviewed-by: Naoumenko, PaulX <paulx.naoumenko@intel.com>
Reviewed-by: Bondier, CedricX <cedricx.bondier@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>