[PORT FROM MAIN] Accelerate the thumbnail extraction when the sync-frame duration is long
BZ: 58036

When the sync frame duration is very long in the test clips, the
thumbnail extration will be very slow. Because the program will
have to decode from the previous sync-frame to the desired frame.
If end user presses the exit button before the thumbnail
extraction finishes, sometimes ANR will happen.

In this patch, if the desized frame's timestamp is 3secs larger
than the previous sync-frame, we will choose to use the sync-frame
as the thumbnail.

Change-Id: I2178e62e2294026179013751d21a2bd73d20f292
Orig-Change-Id: Ic1b7efe5a090908d4cb771a7497457888246657b
Signed-off-by: wfeng6 <wei.feng@intel.com>
Signed-off-by: Gu, Wangyi <wangyi.gu@intel.com>
Reviewed-on: http://android.intel.com:8080/121767
Reviewed-by: Liang, Dan <dan.liang@intel.com>
Reviewed-by: Ding, Haitao <haitao.ding@intel.com>
Tested-by: Ding, Haitao <haitao.ding@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>