Accelerate the thumbnail extraction when the sync-frame duration is long
BZ: 58036

When the sync frame duration is very long in the test clips, the
thumbnail extration will be very slow. Because the program will
have to decode from the previous sync-frame to the desired frame.
If end user presses the exit button before the thumbnail
extraction finishes, sometimes ANR will happen.

In this patch, if the desized frame's timestamp is 3secs larger
than the previous sync-frame, we will choose to use the sync-frame
as the thumbnail.

Change-Id: Ic1b7efe5a090908d4cb771a7497457888246657b
Signed-off-by: wfeng6 <wei.feng@intel.com>
Reviewed-on: http://android.intel.com:8080/75049
Reviewed-by: lab_cactus <lab_cactus@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Gu, Wangyi <wangyi.gu@intel.com>
Tested-by: Ding, Haitao <haitao.ding@intel.com>
Tested-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: Tong, BoX <box.tong@intel.com>