telephony: Fix issue in PhoneNumberUtils.compareLoosely
BZ: 64234

If the number of matched digits is greater than or
equal to MIN_MATCH and if one of the input strings
index has reached the lower limit,
PhoneNumberUtils.compareLoosely should return
true. Currently, it doesn't return true as it is
expecting the index to be < 0 which will not happen
for the following case(country code+number without
"+" prefix) and number without country code)

This patch fixes the issue by checking the index <= 0.

Change-Id: I1d33ad2ba3641fcb78ed2da5a0bb0c4d152ab0ac
Orig-Change-Id: Ic76294ce69598c6bda2e60aeaebac23db233b457
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/81027
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>