ignore NORMAL mode
BZ: 63505

Root Cause of issue is linked to timing problem to stop
the threads. Because the issue is not reproducible on
reference phone, one of possible fixes is to find different
between our code sources and reference phones code sources.
According comments in code source, there is issue on
Samsung tuna and herring boards (i.e. on reference phones).
They are defined a workaround and this patch purpose extended
the Samsung workaround to Intel products..

Change-Id: I73b6da487138a636d664be004307e819b2bbe768
Origin-Change-Id: If4d311b7e8efe00020c38df76f20e1f37987b5cf
Origin-Change-Id: Ibf99954cfa20db23642f22eb7ed80280fa6754d5
Signed-off-by: Andrei V. FOMITCHEV <andreix.fomitchev@intel.com>
Reviewed-on: http://android.intel.com:8080/78359
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>