[PORT FROM MAIN]JAVA Crash in Google Map when run Monkey test
BZ: 61569

Rootcause: If the device have hardware key,When the Monkey test
send the KEY_MENU event quickly,Before the parent window dispatch
it's token to it's view. If the KEY_MENU event come.The sub window
will show, and it will don't get the correct token from parent window.
Then it crash.

Solution: This is a workround fix.Before show the sub window.we do a
check first.if the token got from parent is incorrect,stop the process.

Change-Id: I1df7f3da03c3bc430559386e9d208243f26bc32a
Signed-off-by: jli119X <jianpingx.li@intel.com>
Reviewed-on: http://android.intel.com:8080/76018
Reviewed-by: lab_cactus <lab_cactus@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Li, XiaojingX <xiaojingx.li@intel.com>
Tested-by: Li, XiaojingX <xiaojingx.li@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>