Telephony: The settings values aren't taken in account correctly.
BZ: 66630

Just after a phone flash, the phone receives and
displays the Ambert Alert SMS CB, even if the
"Show Amber Alert" is disabled in the setting
of the CellBroadcastReceiver application by
default. This default value is to false in the
preference.xml file of the CellBroadcastReceiver
application in the vendor/intel/common/overlays
folder.

This patch updates the preference
TwoStatePreference.java file of the telephony
framework, with which after a phone flash the
Amber Alert for SMS CB is not received and
displayed, when the "Show Amber Alert" setting
disabled. Also With this patch the xml files are
correctly initialized after a phone flash.
For more information see the comments 13 to 25 of
the BZ 66630.

Change-Id: Idf8dc8b0b8e1376dde80051638f01612e5b84503
Orig-Change-Id: I16bad467b1ec1598d24f9d5fd045cd52ff8ac83f
Signed-off-by: Delaude Emmanuel <emmanuelx.delaude@intel.com>
Reviewed-on: http://android.intel.com:8080/81028
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>