[PORT FROM MAIN] Telephony: The settings values aren't taken in account correctly.
BZ: 66630

Just after a phone flash, the phone receives and
displays the Ambert Alert SMS CB, even if the
"Show Amber Alert" is disabled in the setting
of the CellBroadcastReceiver application by
default. This default value is to false in the
preference.xml file of the CellBroadcastReceiver
application in the vendor/intel/common/overlays
folder.

This patch updates the preference
TwoStatePreference.java file of the telephony
framework, with which after a phone flash the
Amber Alert for SMS CB is not received and
displayed, when the "Show Amber Alert" setting
disabled. Also With this patch the xml files are
correctly initialized after a phone flash.
For more information see the comments 13 to 25 of
the BZ 66630.

Severity: medium

Change-Id: Ibb6273b39dbcd6fe9354a720dd5374fab19a4ae9
Orig-Change-Id: I9bc17faea5e6e9f3a9c3ce130834f6f82b9d06ca
Category: device enablement
Domain: Telephony-Modem-Intel
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Delaude Emmanuel <emmanuelx.delaude@intel.com>