[PORT FROM MAIN] wps: update the network configuration when connecting using WPS_PBC
BZ: 62955

When the framework uses wps_pbc, it leaves the wpa_supplicant.conf's
update to the supplicant. The issue is that some fields are not
written by supplicant such as priority and bgscan.
So, unlike the connect function, the startWpsPbc doesn't add these
missing fields by itself.
This patches calls selectNetwork, which calculates the new priorities
and updates the configuration file. The selected network is the newest
one in the configurations list after mWifiNative.startWpsPbc returns.
We should have only one new network.
And the select command doesn't actually interrupt or diturb the already
ongoing wps connection done, but just updates the conf.
This is an example of the new conf file (the wps ssid is linksys_jga)
network={
	ssid="Guest"
	key_mgmt=NONE
	priority=3
	bgscan="learn:10:-74:60"
}

network={
	ssid="linksys_jga"
	psk="widiwidi"
	proto=WPA
	key_mgmt=WPA-PSK
	pairwise=CCMP
	auth_alg=OPEN
	priority=4
	bgscan="learn:10:-74:60"
}

Change-Id: I777f14570663c518f4f83adcbf4388b37efa072f
Signed-off-by: Maher Khenissi <maherx.khenissi@intel.com>
Reviewed-on: http://android.intel.com:8080/80612
Reviewed-by: Garcia, JeremieX <jeremiex.garcia@intel.com>
Reviewed-by: Trivelly, Jean <jean.trivelly@intel.com>
Reviewed-by: De Cesco, Jonathan <jonathan.de.cesco@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>