DebugFeature: dump more info to ease ANR debugging.
BZ: 64066

Dump the dispatch history of the main looper of the app at fault.
Dump the all the messages still in the message queue.
Dump the sync barrier info to help debug vsync related ANR.
Dump the last inpu event, the current and the pending ones.
Dump the dispatch history of the WindowMananger threads.
Turn on App<->IME interaction logs.

Note
1. History of app looper is also saved into dropbox besides aplog.
2. History of WMS will be printed in dumpstate.

Change-Id: Iaa7a8ea9ada55e1aa967e933496536f2308fb82e
Signed-off-by: Guobin Zhang <guobin.zhang@intel.com>
Reviewed-on: http://android.intel.com:8080/76081
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Li, XiaojingX <xiaojingx.li@intel.com>
Tested-by: Li, XiaojingX <xiaojingx.li@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>