Revert "NumberPicker should adjust min and max when displayed values are set." (a.k.a. Santa is back)
BZ: 71419

This reverted change was adjusting the min and max values for the NumberPicker
which is not desirable since it changes behavior and it will be possible for
an app that works on the current platform to crash on an older one. Also the
adjustment was not implemented correctly.

Updated the documentation to clarify the reltionship between the min value,
max value, and the displayed values array.

Bug:7518172

This reverts commit a1410e6789ce72bc423793315a51aea8b6bad6c7

Change-Id: I109f1b1f54c1e609941243cabab9241871b6b12b
Reviewed-on: http://android.intel.com:8080/77317
Reviewed-by: Liu, Shuo <shuo.liu@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Li, XiaojingX <xiaojingx.li@intel.com>
Tested-by: Li, XiaojingX <xiaojingx.li@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>