[PORT FROM MAIN]TOMBSTONE in com.android.browser
BZ: 64237

The WebView obj could be destroyed when Tab is destroyed. Which
is weird because there are reference to it in HTML5VideoViewProxy
obj.

We release the HTML5VideoViewProxy reference in WebView destroy
function here.

And add another workaround to prevent HTML5VideoViewProxy calling
native once WebView destroy operation done.

NOTE: we can't set mWebView in HTML5VideoViewProxy to NULL directly
because it's reference everywhere in HTML5VideoViewProxy and that
could trigger NULL exception. We set another boolen member if Webview
is destroyed and use it in HTML5VideoViewProxy.VideoPlayer.setBaseLayer.

Change-Id: Ic3c4b08945fa353cb658b1dab4456c6901cb4083
Signed-off-by: binx.xu <binx.xu@intel.com>
Reviewed-on: http://android.intel.com:8080/75455
Reviewed-by: lab_cactus <lab_cactus@intel.com>
Reviewed-by: Yin, Fengwei <fengwei.yin@intel.com>
Reviewed-by: Li, XiaojingX <xiaojingx.li@intel.com>
Tested-by: Li, XiaojingX <xiaojingx.li@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>