[PORT FROM MAIN] Telephony: Add IOemTelephony and Oem specific constants
BZ: 25540

Currently, IOemTelephony.aidl resides
inside the device/intel/apps/OemTelephony folder.
Due to this, applications and frameworks are
not able to import this.

There is a requirement for IOemTelephony
to be used in both framework and applications.

With this patch, IOemTelephony interface file and
also OEM specific constants are moved from device
specific folders to the frameworks folder.

Category: Device enablement
Domain: Telephony-base
Origin: Internal
Upstream-Candidate: No, proprietary

Change-Id: I8a31919a2546e5315e5388754fadb6eac59ab7f7
Orig-Change-Id: I28e5487ae86035f7dd08e5f09ae043cef5216277
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>