[PORT FROM MAIN] telephony: Add IOemTelephony and Oem specific constants
BZ: 25540

Currently, IOemTelephony.aidl resides
inside the device/intel/apps/OemTelephony folder.
Due to this, applications and frameworks are
not able to import this.

There is a requirement for IOemTelephony
to be used in both framework and applications.

With this patch, IOemTelephony interface file and
also OEM specific constants are moved from device
specific folders to the frameworks folder.

Change-Id: Ifa9808304535892dee674e4b2ad33656c057c9f4
Orig-Change-Id: I1894dda20e2becf7f8c82023f9cf7f675f56f5c1
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/78820
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>