Enable ALS to control LED when Automatic brightness disabled
BZ: 20474

If Automatic Brightness disabled, LED is always OFF. The purpose was to
save the power from Keypad LED, however the user could not find the LED
location in dark/dim environment.

After the discussion, we decide not to save the power in this way, and
it comes to be a bug that we are fixing to have same behavior when
Automatic brightness enabled.

Change-Id: I093cb285874655079c0948c76770a28333958a28
Orig-Change-Id: Ifa7520ccab0fe045a843356b6ae8152f472c25ed
Signed-off-by: Leo Yan <leo.yan@intel.com>
Reviewed-on: http://android.intel.com:8080/76948
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Wang, Zhifeng <zhifeng.wang@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>