Secure a handful of framework broadcasts
BZ: 73011

Applications able to set HTTP proxy
ID: ANDROID-7622253
Severity: high
Affected versions: 3.1 and later

On Android versions 3.1 and later, applications were inadvertently allowed to set a device-wide HTTP proxy. This issue is publicly known and is in use. A CTS test for this issue will be provided.

At least one of these was getting spoofed in the wild.
We don't want to go too far this late in the game, but this is a good start.

bug:7622253

Change-Id: If6b374e8670d83bc3fcbc0032ea3352798d04b63
Signed-off-by: Robert Greenwalt <rgreenwalt@google.com>
Signed-off-by: Xiaokang Qin <xiaokang.qin@intel.com>
Reviewed-on: http://android.intel.com:8080/83683
Reviewed-by: Yin, Fengwei <fengwei.yin@intel.com>
Reviewed-by: Li, XiaojingX <xiaojingx.li@intel.com>
Tested-by: Li, XiaojingX <xiaojingx.li@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>