[PORT FROM MAIN] Telephony: Enabling Network Personalization PUK
BZ: 18871

Network personalisation is the feature were a SIM is restricted
to be used in a specific operator networks only. If an unsupported
SIM is used in the device, the user is asked to enter an unlock
code, so that the SIM can be used in the network. If the user enters
wrong PIN for a preprogrammed number of times, the device is PUK
locked and needs to be taken to the customer care. Android by default
supports entering the PIN, but doesn't inform the user if the device
is PUK locked.

This patch introduces the required enumerations, and INTENTS to be
used in the framework, which can inform the NETWORK PUK lock to
keyguard and lock screen for a suitable display.

Change-Id: I14f1096e56ac0b5a38feec6836ecf53225a74b8f
Orig-Change-Id: I8362dfcaecb67ea8f65ba21bc370b8b9df0bec55
Category: aosp improvement
Domain: Telephony-base
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>