[PORT FROM MAIN]EM/PowerUI: Screen will be turned on at low battery status
BZ: 52107
[R4 MAINLINE]http://android.intel.com:8080/#/c/64115/
When Screen is off then low battery popup is not visible to
user. With this patch we make the screen turn on just before
popup and it will remain in the state for 3 sec, after that
default turn off of the screen will take place. One more
condition is added, if sound is in off state then vibra will
be turned on for 0.5 sec. Tested the code at 14% popup: screen
turn on, made some sound and vibrated similar to what we used
get when popup comes during screen on state.

Change-Id: If665f47f27d933f59ce5f5895dac91091e7f2fdf
Signed-off-by: vinaygox <vinayx.govindaiah@intel.com>
Reviewed-on: http://android.intel.com:8080/79147
Reviewed-by: Chaurasia, Avinash K <avinash.k.chaurasia@intel.com>
Reviewed-by: R, Durgadoss <durgadoss.r@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Mitra, DebalinaX <debalinax.mitra@intel.com>
Tested-by: Kallappa Manjanna, MadhukumarX <madhukumarx.kallappa.manjanna@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>