[PORT FROM MAIN][All] Inconsistency in Audio Behaviour when transeferring from Bluetooth to Wired Headset.
BZ: 68511

While BT is ON, when the wired headset is slowly inserted, the transistion from headphone
to headset is causing this issue. In this transistion, the headset gets connected first
and then the headphone gets disconnected. Since the last operation was disconnection the
output device changes to BT. This fix will check whether either of the device(headset or
headphone) is connected before changing the device to BT.

Change-Id: I49902b2dd93750a5f6aee8c304db66b9d380e359
Orig-Change-Id: I0b3a62b3917d869bfede428da1fe4baf239c0bf3
Signed-off-by: Mazhar <mazharx.sameullah@intel.com>
Reviewed-on: http://android.intel.com:8080/123263
Reviewed-by: M, Arulselvan <arulselvan.m@intel.com>
Tested-by: M, Arulselvan <arulselvan.m@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>