[PORT FROM MAIN][All] Inconsistency in Audio Behaviour when transeferring from Bluetooth to Wired Headset.
BZ: 68511

While BT is ON, when the wired headset is slowly inserted, the transistion from headphone
to headset is causing this issue. In this transistion, the headset gets connected first
and then the headphone gets disconnected. Since the last operation was disconnection the
output device changes to BT. This fix will check whether either of the device(headset or
headphone) is connected before changing the device to BT.

Category: aosp improvement
Domain: AUDIO_MEDIA-Framework-Media
Origin: Internal
Upstream-Candidate: yes

Change-Id: I3c7e1676b6deecf1004e1c46ed1b909ea05bd9f2
Orig-Change-Id: I49902b2dd93750a5f6aee8c304db66b9d380e359
Signed-off-by: Mazhar <mazharx.sameullah@intel.com>