Wifi: Binding/unbinding to the modem manager service for wifi on and off
BZ: 71562 85870

DO NOT UPSTREAM

According to the new modem management policy, WifiService has to
bind to the ModemManager when wifi is switched on and unbind from it when wifi is switched off.
When WifiService binds to modemmanager,
it requests the modem resource and prevents the modem from being switched off.

Change-Id: Idcfb67131bb61b8232bdf59ded0324eb3afc5009
Signed-off-by: Paul Naoumenko <paulx.naoumenko@intel.com>
Reviewed-on: http://android.intel.com:8080/85051
Reviewed-by: Dupras, HugoX <hugox.dupras@intel.com>
Reviewed-by: Trivelly, Jean <jean.trivelly@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>