[PORT FROM MAIN]: frameworks: Report right battery status on lock screen.
BZ: 55683

Currently, the UI shows 'charging' in lock screen though the charging
has been stopped due to some reason. This is because the Android UI
checks only if the charger is plugged in. But it should actually check
the battery status also to find out if the device is actually 'charging'.
This patch is to show 'plugged in, not charging' if the device is not
charging while plugged in.

Original main patch:
http://android.intel.com:8080/#/c/65734/

Change-Id: I77f560f9ae5b53f0e88a47beaa4b07ffbad21411
Signed-off-by: Saranya Gopal <saranya.gopal@intel.com>
Reviewed-on: http://android.intel.com:8080/82218
Reviewed-by: Pallala, Ramakrishna <ramakrishna.pallala@intel.com>
Reviewed-by: G S, RajkumarX <rajkumarx.g.s@intel.com>
Reviewed-by: Govindaiah, VinayX <vinayx.govindaiah@intel.com>
Reviewed-by: Jena, TapanX <tapanx.jena@intel.com>
Tested-by: Kallappa Manjanna, MadhukumarX <madhukumarx.kallappa.manjanna@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>