[PORT FROM MAIN] The input method dialog can't click after rotate.
BZ: 75266

The root cause is the same as bug 62461. But the steps of reproduce
in this issue will change oldFocus to null.So the fix in bug 62461
can't cover this issue.So add a new condition to cover it.
Category: aosp improvement
Domain: AOSP-Framework-GUI
Severity: medium
Origin: internal
Upstream-Candidate: yes

Orig-Change-Id: Id319755ddd2ffe82203373a7d497ac80c0d37bc2

Change-Id: I3a902421ea3bb18eda845ed0dd1b71839a914090
Signed-off-by: jli119X <jianpingx.li@intel.com>