WiFi: Update RSSI value once associated with the AP
BZ: 74271

RSSI value is reset to MIN_RSSI before connection to an AP.
This value was not updated during connection/authentification.
Then, the displayed RSSI value was MIN_RSSI until the first RSSI poll scan.
This patch fixes this issue by updating RSSI value once associated to the AP
so as to early have a correct value.

Change-Id: I54dc2f07b994fb69ccb3de2c68472d38414eeb10
Signed-off-by: Loic Poulain <lpoulaix>
Reviewed-on: http://android.intel.com:8080/84976
Reviewed-by: Poulain, LoicX <loicx.poulain@intel.com>
Reviewed-by: De Cesco, Jonathan <jonathan.de.cesco@intel.com>
Reviewed-by: Garcia, JeremieX <jeremiex.garcia@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Mauger, MatthieuX <matthieux.mauger@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>