There is a hidden issue about overflow
BZ: 77505

Rootcause: There is a hidden issue about overflow. When app want to
create it's drwaing cache. if it's width or height is large. The size
of the bitmap will be width*height*4. The size will beyond the range
of integer.So convert int to long first to avoid overflow

Change-Id: I868652e8bc6c96fa83c752526ef189418f6653b7
Signed-off-by: jli119X <jianpingx.li@intel.com>
Reviewed-on: http://android.intel.com:8080/85271
Reviewed-by: Zhang, Hongyu <hongyu.zhang@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Li, XiaojingX <xiaojingx.li@intel.com>
Tested-by: Li, XiaojingX <xiaojingx.li@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>