[PORT FROM MAIN] There is a hidden issue about overflow
BZ: 77505

Rootcause: There is a hidden issue about overflow. When app want to
create it's drwaing cache. if it's width or height is large. The size
of the bitmap will be width*height*4. The size will beyond the range
of integer.So convert int to long first to avoid overflow

Category: AOSP Improvement
Domain: AOSP-Framework-GUI
Oringin: Internal
Upstream-Candicate: no, needs rework

Orig-Change-Id: I868652e8bc6c96fa83c752526ef189418f6653b7
Change-Id: I250d2180d046f64077014b7f66aeca86c7f020ee
Signed-off-by: jli119X <jianpingx.li@intel.com>