[PORT FROM MAIN] There is a hidden issue about overflow
BZ: 77505

Rootcause: There is a hidden issue about overflow. When app want to
create it's drwaing cache. if it's width or height is large. The size
of the bitmap will be width*height*4. The size will beyond the range
of integer.So convert int to long first to avoid overflow
Category: aosp improvement
Domain: AOSP-Framework-GUI
Severity: medium
Origin: internal
Upstream-Candidate: yes

Orig-Change-Id: I868652e8bc6c96fa83c752526ef189418f6653b7

Change-Id: I57a8d031d30a19367a36851c1a9a0546ba61aac3
Signed-off-by: jli119X <jianpingx.li@intel.com>