WLAN: Temporary SleepPolicy integration on JBMR1 [DO NOT UPSTREAM]
BZ: 78692

This is a squash of below ported patches + few adaptation for JBMR1:
This patch will be reworked for brodcaom sleepPolicy integration using PNO

1) WLAN: SleepPolicy / Periodic scan integration
	BZ: 53958

- Add periodic scan Command in Native Layer
- When periodic scan should be enabled, get the first 16 remembered
network and enable periodic scan
- Update Wifi state machine to manage periodic scan start/stop command
& automatic restart after connection lost if previously enabled
- Manage transition when Display already off when unplug and/or disconnect

Orig-Change-Id: I94968fa9bcce6342b01eb8a33057935deb97f338

2) WLAN: Align framework on supplicant periodic scan rework
	BZ: 56506

Periodic scan should only be enable when supplicant is not connected.
However, when DUT is within the range limit, device is potentially
disconnecting (beacon loss) and reconnecting quickly. As a side effect,
wpa_supplicant is often switching states and many wpa_supplicant state
messages are triggered/sent.

In the meantime, framework message processing is not based on real time
constraints mechanism and unfortunately sometime adding delay in message
handling.

In the case of periodic scan enable/disable messages, this delay could
impact wpa_supplicant state machine if real supplicant is not SCANNING state.

Another side effect is that periodic scan is not re-enable properly upon
re-connection as periodic scan is handled by Android framework
level. There is a misalignment between wpa_supplicant and framework and
device can only recover when display is back on

A rework was done at supplicant level to handle periodic scan offload
suspend/resume when DUT connect/disconnect.

This fix align framework on supplicant changes.

Orig-Change-Id: I93aafc4c19eda2d37d25c08a8e293b6bc256e03a

3) WIFI:  Change automatic periodic scan stop algorithm
	BZ: 35959

when DUT is re-connecting due to AP is found from periodic scan,
periodic scan stop occurs too late in current implementation (once
connection is complete). As a consequence, some trouble during
wifi connection could occurs.

This fix consist in moving periodic scan stop to wpa_supplicant.
As it the supplicant that start the connection, it's easy for it
to stop periodic scan stop before processing new connection.

Orig-Change-Id: Ia370b26ada9b35a62584becc1440c6201e872727

4) wifi: Unexpected periodic scan command initiated by framework makes a WiFi f/w recovery
	BZ: 64587

DO NOT UPSTREAM

Sometimes, unexpected periodic scan is launched by the framewrok because of
a wrong Display State : mScreenOff variable indicates true whereas Display is
ON. This can be due to latency in the Android Java Intent management.
As a consequence, a periodic scan command is initiated by the framewrok down to the
TI f/w. Unlucky, if a normal scan is already on going, the periodic scan generates
a f/w recovery without any restore of this periodic scan command after. Then, no more
scan is triggered and user must switch OFF/ON WiFi to recover.

Fix is to add robustness to ensure that Screen is well OFF before start the periodic scan.
PowerManager.isScreenOn API allows to retrive the state of DISPLAY : it retruns true if DISPLAY is ON,
else returns FALSE.
So, add this test as a condition for periodic scan enabling.

Orig-Change-Id: I5dfd1a74e7b4763f360ce5c9d62a970906510cda

5) wifi: NullPointerException if SSID is null
	BZ: 72699

When starting WPS PBC, WifiConfigStore::mConfiguredNetworks could
contain NULL SSID before the WPS Association terminates, because
in startWpsPbc, we have:
        if (mWifiNative.startWpsPbc(config.BSSID)) {
            loadConfiguredNetworks();
In our case, the NullPointerException occurs after the reception
of CMD_START_PERIODIC_SCAN. We had added robustness to startWpsPbc
to handle NULL SSID, and this patch enforces robustness in this
particular case.

Orig-Change-Id: I76fe618681007a8fc45bed0d5ec9d860df413363
                    ------------------

Change-Id: I805785f3c0742bd1d7422dcc0c80115438cb99f8
Signed-off-by: Honore Tricot <honorex.tricot@intel.com>
Reviewed-on: http://android.intel.com:8080/85835
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Reviewed-by: Fiat, Christophe <christophe.fiat@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: Trivelly, Jean <jean.trivelly@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>