core: Move new resources to end of the list
BZ: 80336

DO NOT UPSTREAM

Resources are introduced in the middle thus causing
all following resources to be assigned new resource
numbers. This has broken the backward compatibility
with old applications. Due to this, options are
shown even when it is not supposed to be shown.

With this patch, newly added resources are moved to the
end of the list of resources to maintain backward
compatibility.

Change-Id: I866ce4f29ebc6d50963417e185e510659437e829
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/86966
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: Roques, CyrilX <cyrilx.roques@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>