[PORT FROM R4.1] Audio-MW: Silent mode changed automatically after touch into volume setting and return with do nothing.
BZ: 73920

In setStreamVolume() of AudioService.java, if volume is muted and vibrator is
set then it sets VIBRATE mode by default. Due to which SILENT mode is changed
automatically to VIBRATE mode after touch into volume setting and return doing
nothing. This fix gets the ringer mode and if it is SILENT mode then does not
change it further.

Change-Id: I44dfe2f8b73577fd7b48afeb10093926d4685c6d
Signed-off-by: gbhakteX <gurudattax.bhakte@intel.com>
Reviewed-on: http://android.intel.com:8080/87301
Reviewed-by: Kandasamy, Muthukumar <muthukumar.kandasamy@intel.com>
Reviewed-by: Sameullah, MazharX <mazharx.sameullah@intel.com>
Reviewed-by: Sikkandar D, Madar <madar.sikkandar.d@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>