EM/Charger: graceful shutdown when low battery with charger connected
BZ: 63823

When SDP connected and the battery is at 0% capacity the phone dosen't
shutdown. The shutdown action at 0% is initiated from the kernel space.
This has an effect on the recalculation of FG_ALRT soc threshold wherein
the interrupts triggering are missed for charging/discharging.
So, a seperate thread is spawned from user space at 0% which handles the
shutdown event at low battery and charger connected condition and does not
effect the device booting.

Change-Id: Id69555826b5d9f713b01e749238712cefbd77de0
Signed-off-by: Naveen Rapaka <naveen.rapaka@intel.com>
Reviewed-on: http://android.intel.com:8080/91194
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Gopal, Saranya <saranya.gopal@intel.com>
Reviewed-by: R, Durgadoss <durgadoss.r@intel.com>
Reviewed-by: Pallala, Ramakrishna <ramakrishna.pallala@intel.com>
Reviewed-by: Beare, Bruce J <bruce.j.beare@intel.com>
Tested-by: Kallappa Manjanna, MadhukumarX <madhukumarx.kallappa.manjanna@intel.com>
Reviewed-by: Kundella, SreedharX <sreedharx.kundella@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>