Check the InputManager service availability
BZ: 86288

If the InputManager service is not available during an 'input'
command, the command will crash with a
java.lang.NullPointerException. As seen in the bz: 74972, it can be
confusing for bug analyses as the Exception will also dump JavaBinder
and ServiceManager traces.
This patch adds some checks in order to deal with such cases. The
'input' command will now exit cleanly and report an 'InputManager
service is not available' error to the user.

Change-Id: I550217f8c2f4f8be55ddd895c380d016355fce9e
Signed-off-by: Clement Calmels <clement.calmels@intel.com>
Reviewed-on: http://android.intel.com:8080/91554
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Tested-by: Shen, JianchunX <jianchunx.shen@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>