[PORT FROM MAIN] Check the InputManager service availability
BZ: 86288

If the InputManager service is not available during an 'input'
command, the command will crash with a
java.lang.NullPointerException. As seen in the bz: 74972, it can be
confusing for bug analyses as the Exception will also dump JavaBinder
and ServiceManager traces.
This patch adds some checks in order to deal with such cases. The
'input' command will now exit cleanly and report an 'InputManager
service is not available' error to the user.

Category: aosp improvement
Domain: AOSP-Framework-OS
Severity: medium
Origin: internal
Upstream-Candidate: yes

Change-Id: Ia85a1e502c280288a8089f4748bcdb0c186e9f60
Orig-Change-Id: I550217f8c2f4f8be55ddd895c380d016355fce9e
Signed-off-by: Clement Calmels <clement.calmels@intel.com>